Reorder all functions in Store
