+++ src/data/mod.rs
@@ -4,3 +4,4 @@ pub mod filesystem;
 pub mod preferences;
 pub mod search;
 pub mod store;
+mod faked_models;
